Fixed dropTables;

+++ dbBackend/src/main/resources/cz/muni/fi/pv168/db_backend/backend/dropTables.sql
@@ -1,4 +1,3 @@
-DROP TABLE agents;
-DROP TABLE assignments;
-DROP TABLE missions;
-DROP TABLE sp_powers;
+DROP TABLE agents if EXISTS;
+DROP TABLE assignments if EXISTS;
+DROP TABLE missions if EXISTS;
